Fix: unary operator without element to operate on inside grouping causing bad output
Fixes #106

This modification fixes cases like `(hat)` where a UNARY operator is inside grouping symbols.
Previously it would output something equivalent to `(hathat)`, which isn't ideal.  For
mi and mo UNARY symbols, the current behavior will remain, but for accent and font-change symbols,
a blank element will be inserted for the unary to operate on.